Moving schema files under utils.
