+++ google/gmp/example_dags/sequential_erf_uploader_to_bq_dag.py
@@ -22,7 +22,7 @@ from datetime import datetime
 from datetime import timedelta
 from airflow import DAG
 from airflow import models
-from schema import Entity_Schema_Lookup
+from google.gmp.utils.schema import Entity_Schema_Lookup
 from google.gmp.operators.gmp_dv360_operator import DisplayVideo360ERFUploadBqOperator
