+++ google/gmp/utils/schema/CustomAffinity.py
