+++ google/gmp/utils/schema/InventorySource.py
