+++ google/gmp/utils/schema/Partner.py
