+++ google/gmp/utils/schema/Pixel.py
