+++ google/gmp/utils/schema/sdf/v4_2/Campaign.py
